requirements.txt
iris-pex-embedded-python>=3.0.0
pandas>=1.5.3
rdkit>=2022.9.4
xgboost-cpu
scikit-learn>=1.2.1
fastapi>=0.70.0
python-multipart>=0.0.5
requests>=2.28.2
joblib>=1.2.0
sqlalchemy-iris>=0.5.0
uvicorn>=0.15.0
huggingface-hub
sentence-transformers
datamol